+++ moveHelper.js
